fix(controllers): change predicate rate to double type

+++ source/modulo_controllers/src/BaseControllerInterface.cpp
@@ -47,7 +47,7 @@ BaseControllerInterface::on_configure(const rclcpp_lifecycle::State&) {
     predicate_message_.type = modulo_interfaces::msg::PredicateCollection::CONTROLLER;
 
     predicate_timer_ = get_node()->create_wall_timer(
-        std::chrono::nanoseconds(static_cast<int64_t>(1e9 / get_parameter_value<int>("predicate_publishing_rate"))),
+        std::chrono::nanoseconds(static_cast<int64_t>(1e9 / get_parameter_value<double>("predicate_publishing_rate"))),
         [this]() { this->publish_predicates(); });
   }
